# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Website analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analysis]
      summary: List all analyses
      description: Returns a summary of all analyses in the current session
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesListResponse"

  /analyze:
    post:
      operationId: startAnalysis
      tags: [analysis]
      summary: Start a new website analysis
      description: Initiates analysis with questionnaire data and URL or HTML input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAnalysisBody"
      responses:
        "201":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisStarted"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyze/{id}:
    get:
      operationId: getAnalysisReport
      tags: [analysis]
      summary: Get analysis report
      description: Returns the completed analysis report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisReport"
        "404":
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    QuestionnaireData:
      type: object
      properties:
        companyPitch:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        brandName:
          type: ["string", "null"]
        brandVariants:
          type: ["string", "null"]
        subBrands:
          type: ["string", "null"]
        slogans:
          type: ["string", "null"]
        buyerPersonas:
          type: ["string", "null"]
        geographicFocus:
          type: ["string", "null"]
        contentLanguages:
          type: ["string", "null"]
        competitors:
          type: ["string", "null"]
        differentiators:
          type: ["string", "null"]
        influencers:
          type: ["string", "null"]
        socialMedia:
          type: ["string", "null"]
        microsites:
          type: ["string", "null"]
        directories:
          type: ["string", "null"]
        reviewPlatforms:
          type: ["string", "null"]
        seoTools:
          type: ["string", "null"]
        dataSources:
          type: ["string", "null"]
        restrictions:
          type: ["string", "null"]
        strategicPriority:
          type: ["string", "null"]
        kpis:
          type: ["string", "null"]
        weightingPreferences:
          type: ["string", "null"]
        plannedCampaigns:
          type: ["string", "null"]

    AnalysisSummary:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        url:
          type: ["string", "null"]
        mode:
          type: string
          enum: [url, html]
        overallScore:
          type: ["number", "null"]
        progress:
          type: number
        crawledPagesCount:
          type: integer
        startedAt:
          type: string
      required:
        - id
        - status
        - mode
        - progress
        - crawledPagesCount
        - startedAt

    AnalysesListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisSummary"
        total:
          type: integer
      required:
        - items
        - total

    StartAnalysisBody:
      type: object
      properties:
        mode:
          type: string
          enum: [url, html]
        url:
          type: ["string", "null"]
        html:
          type: ["string", "null"]
        questionnaire:
          $ref: "#/components/schemas/QuestionnaireData"
        explicitUrls:
          type: array
          items:
            type: string
          nullable: true
          description: If provided, skip auto-crawl and use these URLs directly
      required:
        - mode

    AnalysisStarted:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
      required:
        - id
        - status

    TechnicalSeoResult:
      type: object
      properties:
        score:
          type: number
        responseTime:
          type: number
        ttfb:
          type: number
        httpStatusCodes:
          type: object
          additionalProperties:
            type: integer
        robotsTxt:
          type: boolean
        sitemapXml:
          type: boolean
        canonicalTags:
          type: object
          properties:
            present:
              type: boolean
            count:
              type: integer
          required: [present, count]
        hreflang:
          type: object
          properties:
            present:
              type: boolean
            languages:
              type: array
              items:
                type: string
            consistent:
              type: boolean
          required: [present, languages, consistent]
        metaTitles:
          type: object
          properties:
            present:
              type: integer
            avgLength:
              type: number
            tooShort:
              type: integer
            tooLong:
              type: integer
            missing:
              type: integer
          required: [present, avgLength, tooShort, tooLong, missing]
        metaDescriptions:
          type: object
          properties:
            present:
              type: integer
            avgLength:
              type: number
            tooShort:
              type: integer
            tooLong:
              type: integer
            missing:
              type: integer
          required: [present, avgLength, tooShort, tooLong, missing]
        imageAltCoverage:
          type: number
        mobileViewport:
          type: boolean
        httpsEnforced:
          type: boolean
      required:
        - score
        - responseTime
        - ttfb
        - robotsTxt
        - sitemapXml
        - canonicalTags
        - hreflang
        - metaTitles
        - metaDescriptions
        - imageAltCoverage
        - mobileViewport
        - httpsEnforced

    SchemaOrgResult:
      type: object
      properties:
        score:
          type: number
        detectedTypes:
          type: array
          items:
            type: string
        missingHighValue:
          type: array
          items:
            type: string
        productSchemaDetails:
          type: object
          additionalProperties:
            type: boolean
        validationErrors:
          type: array
          items:
            type: string
      required:
        - score
        - detectedTypes
        - missingHighValue
        - validationErrors

    HeadingResult:
      type: object
      properties:
        score:
          type: number
        pages:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              h1Count:
                type: integer
              hasHierarchyIssues:
                type: boolean
              headings:
                type: array
                items:
                  type: object
                  properties:
                    level:
                      type: string
                    text:
                      type: string
                  required: [level, text]
            required: [url, h1Count, hasHierarchyIssues, headings]
        keywordInHeadings:
          type: boolean
      required:
        - score
        - pages
        - keywordInHeadings

    ContentRelevanceResult:
      type: object
      properties:
        score:
          type: number
        dimensions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              score:
                type: number
              findings:
                type: array
                items:
                  type: string
            required: [name, score, findings]
      required:
        - score
        - dimensions

    FaqResult:
      type: object
      properties:
        score:
          type: number
        faqItemsFound:
          type: integer
        hasFaqSchema:
          type: boolean
        hasHtmlFaq:
          type: boolean
        qualityAssessment:
          type: ["string", "null"]
      required:
        - score
        - faqItemsFound
        - hasFaqSchema
        - hasHtmlFaq

    LlmQuestion:
      type: object
      properties:
        question:
          type: string
        rating:
          type: number
        gap:
          type: string
      required:
        - question
        - rating
        - gap

    LlmDiscoverabilityResult:
      type: object
      properties:
        score:
          type: number
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LlmQuestion"
      required:
        - score
        - questions

    CompetitorFindings:
      type: object
      properties:
        betterThanYou:
          type: string
        yourAdvantage:
          type: string
        recommendation:
          type: string
      required:
        - betterThanYou
        - yourAdvantage
        - recommendation

    CompetitorScore:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        technicalScore:
          type: number
        schemaScore:
          type: number
        contentScore:
          type: number
        headingScore:
          type: number
        faqScore:
          type: number
        compositeScore:
          type: number
        crawledPagesCount:
          type: integer
        findings:
          oneOf:
            - $ref: "#/components/schemas/CompetitorFindings"
            - type: "null"
      required:
        - name
        - url
        - technicalScore
        - schemaScore
        - contentScore
        - headingScore
        - faqScore
        - compositeScore
        - crawledPagesCount

    CompetitorResult:
      type: object
      properties:
        competitors:
          type: array
          items:
            $ref: "#/components/schemas/CompetitorScore"
      required:
        - competitors

    Recommendation:
      type: object
      properties:
        tier:
          type: string
          enum: [critical, high_leverage, secondary]
        finding:
          type: string
        whyItMatters:
          type: string
        fixInstruction:
          type: string
      required:
        - tier
        - finding
        - whyItMatters
        - fixInstruction

    AnalysisReport:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        url:
          type: ["string", "null"]
        mode:
          type: string
          enum: [url, html]
        overallScore:
          type: ["number", "null"]
        currentModule:
          type: ["string", "null"]
        progress:
          type: number
        technicalSeo:
          oneOf:
            - $ref: "#/components/schemas/TechnicalSeoResult"
            - type: "null"
        schemaOrg:
          oneOf:
            - $ref: "#/components/schemas/SchemaOrgResult"
            - type: "null"
        headingStructure:
          oneOf:
            - $ref: "#/components/schemas/HeadingResult"
            - type: "null"
        contentRelevance:
          oneOf:
            - $ref: "#/components/schemas/ContentRelevanceResult"
            - type: "null"
        faqQuality:
          oneOf:
            - $ref: "#/components/schemas/FaqResult"
            - type: "null"
        llmDiscoverability:
          oneOf:
            - $ref: "#/components/schemas/LlmDiscoverabilityResult"
            - type: "null"
        competitorComparison:
          oneOf:
            - $ref: "#/components/schemas/CompetitorResult"
            - type: "null"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        errors:
          type: array
          items:
            type: string
        crawledPages:
          type: array
          items:
            type: string
      required:
        - id
        - status
        - mode
        - progress
        - recommendations
        - errors
        - crawledPages