# lib/api-zod/src/generated/types/startAnalysisBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionnaireData } from "./questionnaireData";
import type { StartAnalysisBodyMode } from "./startAnalysisBodyMode";

export interface StartAnalysisBody {
  mode: StartAnalysisBodyMode;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  html?: string | null;
  questionnaire?: QuestionnaireData;
  /** If provided, skip auto-crawl and use these URLs directly */
  explicitUrls?: string[] | null;
}